+++ NEWS.md
@@ -1,3 +1,13 @@
+# spsUtil 0.2.1
+
+## New feature
+
+-   New function `strUniquefy`: takes a string or numeric vector and found all duplicated elements and add an index to the end to make sure all elements are unique.
+
+## Minor Change
+
+-   Use faster `.packages(TRUE)` method in `checkNamespace` to find installed packages.
+
 # spsUtil 0.2.0
 
 ## New feature